requirements.txt
# Constitution Analysis Pipeline - Requirements
# Python 3.8+

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# LLM APIs
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
boto3>=1.34.0
botocore>=1.34.0

# Web Search
requests>=2.31.0

# Environment and Configuration
python-dotenv>=1.0.0

# Progress Bar
tqdm>=4.65.0

# Encoding Detection (for utils/encoding_fix.py)
chardet>=5.0.0

# Optional: Jupyter support (for notebooks)
ipykernel>=6.20.0
jupyter>=1.0.0
